fix(orchestrator): Use correct property for scenario scores
The `simulerEquilibrageScoresBasique` and `calculerScoreEquilibrageScenario`
functions were attempting to access student scores using a `SCORE_` prefix
(e.g., `SCORE_COM`) which does not exist on the student object. The correct
property is the scenario name itself (e.g., `COM`).

This commit corrects the property access in both functions to use the
correct scenario name, allowing the score-based calculations to
function as intended.

A new test file `tests/ScenarioScoreFix.test.js` is added to verify
this fix and prevent regressions.
